+++ js/src/OptionToggleAspect.js
@@ -1,5 +1,0 @@
-export function OptionToggleAspect(setOptionSelectedAspect){
-    return {
-        toggle: (wrap) => setOptionSelectedAspect.setOptionSelected(wrap, !wrap.isOptionSelected)
-    }
-}